+++ src/Service/Ses/src/ValueObject/Message.php
@@ -34,10 +34,18 @@ final class Message
     private $headers;
 
     /**
+     * The List of attachments to include in your email. All recipients will receive the same attachments.
+     *
+     * @var Attachment[]|null
+     */
+    private $attachments;
+
+    /**
      * @param array{
      *   Subject: Content|array,
      *   Body: Body|array,
      *   Headers?: null|array<MessageHeader|array>,
+     *   Attachments?: null|array<Attachment|array>,
      * } $input
      */
     public function __construct(array $input)
@@ -45,6 +53,7 @@ final class Message
         $this->subject = isset($input['Subject']) ? Content::create($input['Subject']) : $this->throwException(new InvalidArgument('Missing required field "Subject".'));
         $this->body = isset($input['Body']) ? Body::create($input['Body']) : $this->throwException(new InvalidArgument('Missing required field "Body".'));
         $this->headers = isset($input['Headers']) ? array_map([MessageHeader::class, 'create'], $input['Headers']) : null;
+        $this->attachments = isset($input['Attachments']) ? array_map([Attachment::class, 'create'], $input['Attachments']) : null;
     }
 
     /**
@@ -52,11 +61,20 @@ final class Message
      *   Subject: Content|array,
      *   Body: Body|array,
      *   Headers?: null|array<MessageHeader|array>,
+     *   Attachments?: null|array<Attachment|array>,
      * }|Message $input
      */
     public static function create($input): self
     {
         return $input instanceof self ? $input : new self($input);
+    }
+
+    /**
+     * @return Attachment[]
+     */
+    public function getAttachments(): array
+    {
+        return $this->attachments ?? [];
     }
 
     public function getBody(): Body
@@ -93,6 +111,14 @@ final class Message
             foreach ($v as $listValue) {
                 ++$index;
                 $payload['Headers'][$index] = $listValue->requestBody();
+            }
+        }
+        if (null !== $v = $this->attachments) {
+            $index = -1;
+            $payload['Attachments'] = [];
+            foreach ($v as $listValue) {
+                ++$index;
+                $payload['Attachments'][$index] = $listValue->requestBody();
             }
         }